Add files via upload
Here I conducted an end to end EDA on this bank telemarketing campaign data set to infer knowledge that where bank has to put more effort to improve it's positive response rate. In addition, build a prediction model for the bank to identify in the future the customers who will likely accept or reject the offer.